+++ substrate/frame/fast-unstake/src/mock.rs
@@ -144,7 +144,7 @@ impl pallet_staking::Config for Runtime {
 	type EventListeners = ();
 	type BenchmarkingConfig = pallet_staking::TestBenchmarkingConfig;
 	type WeightInfo = ();
-	type DisablingStrategy = pallet_staking::UpToByzantineThresholdDisablingStrategy;
+	type DisablingStrategy = pallet_staking::UpToThresholdDisablingStrategy;
 }
 
 pub struct BalanceToU256;